# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Medical",
        "description": "Documentation API pour la  gestion Medical",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
        "/api/admissions": {
            "get": {
                "tags": [
                    "Admission & sortie & transfert"
                ],
                "summary": "liste des admissions",
                "description": "Liste des admissions.",
                "operationId": "listTousLesAdmissions",
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, token invalide ou manquant."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admissions/en-cours": {
            "post": {
                "tags": [
                    "Admission & sortie & transfert"
                ],
                "summary": "Admissions en cours",
                "description": "Recuperer l'admissions en cours du patient",
                "operationId": "AdmissionsEnCours",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "description": "ID du patient",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, token invalide ou manquant."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/addAdmission": {
            "post": {
                "tags": [
                    "Admission & sortie & transfert"
                ],
                "summary": "Ajout admission patient",
                "description": "Admission patient.",
                "operationId": "crateAdmissionPatient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Admission"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admission"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/patients/admission/{id}/update": {
            "put": {
                "tags": [
                    "Admission & sortie & transfert"
                ],
                "summary": "Modifier les infos d'une admission",
                "description": "Modifier les infos d'une admission.",
                "operationId": "updateAdmission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de admission à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Admission"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admission mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Infos Admission mis à jour avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Admission"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admission non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admission non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Admission": {
                "title": "Admission",
                "description": "Schéma pour une Admission",
                "properties": {
                    "motif_admission": {
                        "description": "Motif de l'admission",
                        "type": "string",
                        "example": "Maladie contagieuse"
                    },
                    "dossierMedical_id": {
                        "description": "Dossier medical du patient en question",
                        "type": "integer",
                        "example": "4"
                    },
                    "service_id": {
                        "description": "Service",
                        "type": "integer",
                        "example": "2"
                    },
                    "salle_id": {
                        "description": "Salle",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}